Fix TestORMRowTest to work for 32-bit systems
1) TestORMRowTest::testConstructor with data set #0 (array('Foobar',
'20120101020202', 42, 9000.1, true, array(13, 11, 7, 5, 3, 2),
stdClass), true)
Failed asserting that 2147483647 matches expected '20120101020202'.

+++ tests/phpunit/includes/db/TestORMRowTest.php
@@ -164,7 +164,7 @@ class TestORMTable extends ORMTable {
 			'awesome' => 'bool',
 			'stuff' => 'array',
 			'moarstuff' => 'blob',
-			'time' => 'int', // TS_MW
+			'time' => 'float', // TS_MW
 		);
 	}
